refactor: jpa 역할을 infra로 변경
